+++ fullyConnectedNN/ReluActivation.h
@@ -6,11 +6,25 @@
 
 class ReluActivation : public ActivationFunction {
     public:
-        void forward(const Eigen::MatrixXd& input) override {
-            output = input.unaryExpr([](double x) { return std::max(0.0, x); });
+        void forward(const Eigen::MatrixXf& input) override {
+            output = input.unaryExpr([](float x) { return std::max(0.0f, x); });
         }
 
         // add backward code too
+        // void backward(const Eigen::MatrixXf& dvalues) override {
+        //     Eigen::MatrixXf dinputs = dvalues;
+
+        //     for (int i=0; i<dinputs.rows(); i++) {
+        //         for (int j=0; j<dinputs.cols(); j++) {
+        //             if (dinputs(i,j) < 0) dinputs(i,j) =0;
+        //         }
+        //     }
+        // }
+
+
+        
+
+
 
 };
